Fix: Fix default PreferenceFlow on Apple
CFPropertyList has <true>, <false>, <integer> and <real>, so that it
can support bool, int128_t and double inside NSNumber, which we can
map to Boolean, Int and Float in our Preferences (since Int is usually
used a lot more than Long in preferences).

In order to ensure feature parity to prevent surprises across
platforms, we are also dropping support for Long in the default
Preferences implementation for all platforms now, and we no longer
need explicit typing in PreferencesJson which actually doesn't work
for Kotlin/JS anyway since Int and Float are the same during runtime.

Fixes: #19